Move regionToSelection to own file.
